requirements.txt
Flask
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
gunicorn
oauth2client
requests
pymongo
certifi
dnspython
tensorflow
keras
werkzeug
